+++ packages/chart/src/components/index.ts
@@ -1,0 +1,7 @@
+export * from './customized-label';
+export * from './dot';
+export * from './legends';
+export * from './linear-gradient';
+export * from './rect-bar';
+export * from './tick';
+export * from './tooltip-content';
