+++ packages/chart/src/hoc/index.ts
@@ -1,0 +1,1 @@
+export * from './customized-hoc';
