+++ packages/chart/src/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './use-path-bar';
+export * from './use-settings';
