+++ packages/chart/src/icons/index.ts
@@ -1,0 +1,5 @@
+export * from './circle-icon';
+export * from './circle-line-icon';
+export * from './filled-circle-icon';
+export * from './point-icon';
+export * from './stroke-circle-icon';
